Refactoring the User Tests
