+++ frontend/src/Tests/dashboardTests/actions.unit.spec.js
@@ -1,5 +1,5 @@
 import { expect} from 'chai';
-import * as actions from './actions';
+import * as actions from '../../Dashboard/actions';
 
 describe( 'house actions', () => {
   it( 'should create an action when successfully deleting house', () => {
